+++ components/camel-aws/camel-aws2-lambda/src/test/java/org/apache/camel/component/aws2/lambda/integration/LambdaCreateFunctionIT.java
@@ -14,7 +14,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.apache.camel.component.aws2.lambda.localstack;
+package org.apache.camel.component.aws2.lambda.integration;
 
 import java.io.File;
 import java.io.FileInputStream;
@@ -28,11 +28,12 @@ import org.apache.camel.builder.RouteBuilder;
 import org.apache.camel.component.aws2.lambda.Lambda2Constants;
 import org.apache.camel.component.mock.MockEndpoint;
 import org.junit.jupiter.api.Test;
-import software.amazon.awssdk.services.lambda.model.DeleteFunctionResponse;
+import software.amazon.awssdk.services.lambda.model.CreateFunctionResponse;
 
-import static org.junit.Assert.assertTrue;
+import static org.junit.Assert.assertEquals;
+import static org.junit.Assert.assertNotNull;
 
-public class LambdaDeleteFunctionLocalstackTest extends Aws2LambdaBaseTest {
+public class LambdaCreateFunctionIT extends Aws2LambdaBase {
 
     @EndpointInject
     private ProducerTemplate template;
@@ -63,16 +64,12 @@ public class LambdaDeleteFunctionLocalstackTest extends Aws2LambdaBaseTest {
             }
         });
 
-        template.send("direct:deleteFunction", ExchangePattern.InOut, new Processor() {
-            @Override
-            public void process(Exchange exchange) throws Exception {
-
-            }
-        });
-
         assertMockEndpointsSatisfied();
-        DeleteFunctionResponse resp = result.getExchanges().get(0).getIn().getBody(DeleteFunctionResponse.class);
-        assertTrue(resp.sdkHttpResponse().isSuccessful());
+        CreateFunctionResponse resp = result.getExchanges().get(0).getIn().getBody(CreateFunctionResponse.class);
+        assertEquals("GetHelloWithName", resp.functionName());
+        assertEquals("Hello with node.js on Lambda", resp.description());
+        assertNotNull(resp.functionArn());
+        assertNotNull(resp.codeSha256());
     }
 
     @Override
@@ -81,9 +78,7 @@ public class LambdaDeleteFunctionLocalstackTest extends Aws2LambdaBaseTest {
             @Override
             public void configure() throws Exception {
                 String awsEndpoint = "aws2-lambda://GetHelloWithName?operation=createFunction";
-                String deleteFunction = "aws2-lambda://GetHelloWithName?operation=deleteFunction";
-                from("direct:createFunction").to(awsEndpoint);
-                from("direct:deleteFunction").to(deleteFunction).to("mock:result");
+                from("direct:createFunction").to(awsEndpoint).to("mock:result");
             }
         };
     }
