+++ components/camel-aws/camel-aws2-lambda/src/test/java/org/apache/camel/component/aws2/lambda/integration/LambdaDeleteFunctionIT.java
@@ -14,7 +14,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.apache.camel.component.aws2.lambda.localstack;
+package org.apache.camel.component.aws2.lambda.integration;
 
 import java.io.File;
 import java.io.FileInputStream;
@@ -28,12 +28,11 @@ import org.apache.camel.builder.RouteBuilder;
 import org.apache.camel.component.aws2.lambda.Lambda2Constants;
 import org.apache.camel.component.mock.MockEndpoint;
 import org.junit.jupiter.api.Test;
-import software.amazon.awssdk.services.lambda.model.CreateFunctionResponse;
+import software.amazon.awssdk.services.lambda.model.DeleteFunctionResponse;
 
-import static org.junit.Assert.assertEquals;
-import static org.junit.Assert.assertNotNull;
+import static org.junit.Assert.assertTrue;
 
-public class LambdaCreateFunctionLocalstackTest extends Aws2LambdaBaseTest {
+public class LambdaDeleteFunctionIT extends Aws2LambdaBase {
 
     @EndpointInject
     private ProducerTemplate template;
@@ -64,12 +63,16 @@ public class LambdaCreateFunctionLocalstackTest extends Aws2LambdaBaseTest {
             }
         });
 
+        template.send("direct:deleteFunction", ExchangePattern.InOut, new Processor() {
+            @Override
+            public void process(Exchange exchange) throws Exception {
+
+            }
+        });
+
         assertMockEndpointsSatisfied();
-        CreateFunctionResponse resp = result.getExchanges().get(0).getIn().getBody(CreateFunctionResponse.class);
-        assertEquals("GetHelloWithName", resp.functionName());
-        assertEquals("Hello with node.js on Lambda", resp.description());
-        assertNotNull(resp.functionArn());
-        assertNotNull(resp.codeSha256());
+        DeleteFunctionResponse resp = result.getExchanges().get(0).getIn().getBody(DeleteFunctionResponse.class);
+        assertTrue(resp.sdkHttpResponse().isSuccessful());
     }
 
     @Override
@@ -78,7 +81,9 @@ public class LambdaCreateFunctionLocalstackTest extends Aws2LambdaBaseTest {
             @Override
             public void configure() throws Exception {
                 String awsEndpoint = "aws2-lambda://GetHelloWithName?operation=createFunction";
-                from("direct:createFunction").to(awsEndpoint).to("mock:result");
+                String deleteFunction = "aws2-lambda://GetHelloWithName?operation=deleteFunction";
+                from("direct:createFunction").to(awsEndpoint);
+                from("direct:deleteFunction").to(deleteFunction).to("mock:result");
             }
         };
     }
