+++ components/camel-aws/camel-aws2-lambda/src/test/java/org/apache/camel/component/aws2/lambda/integration/LambdaListFunctionsIT.java
@@ -14,7 +14,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.apache.camel.component.aws2.lambda.localstack;
+package org.apache.camel.component.aws2.lambda.integration;
 
 import java.io.File;
 import java.io.FileInputStream;
@@ -28,12 +28,12 @@ import org.apache.camel.builder.RouteBuilder;
 import org.apache.camel.component.aws2.lambda.Lambda2Constants;
 import org.apache.camel.component.mock.MockEndpoint;
 import org.junit.jupiter.api.Test;
-import software.amazon.awssdk.services.lambda.model.PublishVersionResponse;
+import software.amazon.awssdk.services.lambda.model.ListFunctionsResponse;
 
 import static org.junit.Assert.assertEquals;
 import static org.junit.Assert.assertNotNull;
 
-public class LambdaPublishVersionLocalstackTest extends Aws2LambdaBaseTest {
+public class LambdaListFunctionsIT extends Aws2LambdaBase {
 
     @EndpointInject
     private ProducerTemplate template;
@@ -64,18 +64,20 @@ public class LambdaPublishVersionLocalstackTest extends Aws2LambdaBaseTest {
             }
         });
 
-        template.send("direct:publishVersion", ExchangePattern.InOut, new Processor() {
+        template.send("direct:listFunction", ExchangePattern.InOut, new Processor() {
             @Override
             public void process(Exchange exchange) throws Exception {
-                exchange.getIn().setHeader(Lambda2Constants.VERSION_DESCRIPTION, "This is my description");
+
             }
         });
-        assertMockEndpointsSatisfied();
 
-        PublishVersionResponse resp = result.getExchanges().get(0).getMessage().getBody(PublishVersionResponse.class);
-        assertNotNull(resp);
-        assertEquals("GetHelloWithName", resp.functionName());
-        assertEquals("1", resp.version());
+        assertMockEndpointsSatisfied();
+        ListFunctionsResponse resp = result.getExchanges().get(0).getIn().getBody(ListFunctionsResponse.class);
+        assertEquals(1, resp.functions().size());
+        assertEquals("GetHelloWithName", resp.functions().get(0).functionName());
+        assertEquals("Hello with node.js on Lambda", resp.functions().get(0).description());
+        assertNotNull(resp.functions().get(0).functionArn());
+        assertNotNull(resp.functions().get(0).codeSha256());
     }
 
     @Override
@@ -84,9 +86,9 @@ public class LambdaPublishVersionLocalstackTest extends Aws2LambdaBaseTest {
             @Override
             public void configure() throws Exception {
                 String awsEndpoint = "aws2-lambda://GetHelloWithName?operation=createFunction";
-                String publishVersion = "aws2-lambda://GetHelloWithName?operation=publishVersion";
+                String listFunction = "aws2-lambda://GetHelloWithName?operation=listFunctions";
                 from("direct:createFunction").to(awsEndpoint);
-                from("direct:publishVersion").to(publishVersion).to("mock:result");
+                from("direct:listFunction").to(listFunction).to("mock:result");
             }
         };
     }
